refactored towards a separation of concerns/dependencies so python generation code is decoupled from rust core; include it with feature flag ; goal is to have a clean rust core api for dmap that can more cleanly be consumed/wrapped by languages other than just python
added comments for the rust read api

added unit tests for Rust's read api
